Backend/tahanancrafts/requirements.txt
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.7.14
cffi==2.0.0
charset-normalizer==3.4.2
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
defusedxml==0.7.1
dj-database-url==3.0.1
dj-rest-auth==7.0.1
Django==5.2.1
django-allauth==65.10.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==4.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
fonttools==4.60.1
google-auth==2.40.3
gunicorn==23.0.0
idna==3.10
joblib==1.5.2
jwt==1.4.0
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
mysqlclient==2.2.7
numpy==2.3.3
oauthlib==3.3.1
packaging==25.0
pandas==2.3.3
pillow==11.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT == 2.7.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2025.2
requests==2.32.4
requests-oauthlib==2.0.0
resend==2.17.0
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.2
sendgrid==6.12.5
setuptools==80.9.0
six==1.17.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
telesign==3.0.0
telesignenterprise==3.0.0
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
whitenoise==6.11.0
apscheduler==3.10.4
django-apscheduler==0.6.2
easyocr
torch
torchvision
torchaudio
Pillow
opencv-python-headless


